Add first tab auto-initialization
When yterm is opened, no tabs are created by
default, which is not desired. An initial tab should be
auto-created on the current PWD.

This change adds an InitFetch message on the
terminal IPC channel and methods to create and
select an initial tab in TerminalBackService. The
AppCtrl component will send an InitFetch message
when it is mounted to fetch the initial tab and
selected tab state.